add output UserData if set in FinishInfoRequest
This commit adds the ability to output custom UserData in
FinishInfoRequest so when requesting info we can get custom data.

When checking which user belongs to which object, it would be nice to
check this with the UserData field and get the info about the user in
that field. That's why I think this would be a great addition to the
library.