Remove token bloat
Tokens were hex-encoded (as UUIDs) and then base64-encoded.
This might create a false sense of security (as the tokens
look larger than they really are).

Old token example

NTM2ZDAyMDEtODc0Yi00NGE5LTg1MjUtZDcxOTcyYmE5OGRh

which decodes as

536d0201-874b-44a9-8525-d71972ba98da

In the new system, this would be encoded as

U20CAYdLRKmFJdcZcrqY2g

We have also removed the '==' padding from the right.